introduce "bit rename"
