+++ scopes/component/deprecation/deprecate-cmd.ts
@@ -1,5 +1,8 @@
 import chalk from 'chalk';
+import { Workspace } from '@teambit/workspace';
 import { Command, CommandOptions } from '@teambit/cli';
+import { deprecateMany, undeprecateMany } from '@teambit/legacy/dist/scope/component-ops/components-deprecation';
+import { BitIds } from '@teambit/legacy/dist/bit-id';
 import { DeprecationMain } from './deprecation.main.runtime';
 
 export class DeprecateCmd implements Command {
@@ -13,13 +16,24 @@ export class DeprecateCmd implements Command {
   migration = true;
   remoteOp = true;
 
-  constructor(private deprecation: DeprecationMain) {}
+  constructor(private deprecation: DeprecationMain, private workspace: Workspace) {}
 
   async report([id]: [string], { newId }: { newId?: string }): Promise<string> {
-    const result = await this.deprecation.deprecate(id, newId);
+    const result = await this.deprecate(id, newId);
     if (result) {
       return chalk.green(`the component "${id}" has been deprecated successfully`);
     }
     return chalk.bold(`the component "${id}" is already deprecated. no changes have been made`);
+  }
+
+  private async deprecate(id: string, newId?: string): Promise<boolean> {
+    if (this.workspace.isLegacy) {
+      const bitId = this.workspace.consumer.getParsedId(id);
+      await deprecateMany(this.workspace.consumer.scope, new BitIds(bitId));
+      return true;
+    }
+    const componentId = await this.workspace.resolveComponentId(id);
+    const newComponentId = newId ? await this.workspace.resolveComponentId(newId) : undefined;
+    return this.deprecation.deprecate(componentId, newComponentId);
   }
 }
