+++ scopes/component/forking/index.ts
@@ -1,6 +1,5 @@
 import { ForkingAspect } from './forking.aspect';
 
-export type { ForkingMain } from './forking.main.runtime';
-export { ForkAspectConfig } from './forker';
+export type { ForkingMain, ForkAspectConfig } from './forking.main.runtime';
 export default ForkingAspect;
 export { ForkingAspect };
